Add confirmation dialog for when deleting map

+++ src/app/components/users-panel/users-panel.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import {GeoDataService} from '../../services/geo-data.service';
 import {ProjectsService} from '../../services/projects.service';
+import {ModalService} from '../../services/modal.service';
 import {IProjectUser} from '../../models/project-user';
 import {FormGroup, FormControl} from '@angular/forms';
 import {Project} from '../../models/models';
@@ -16,7 +16,7 @@ export class UsersPanelComponent implements OnInit {
   addUserForm: FormGroup;
   activeProject: Project;
 
-  constructor(private projectsService: ProjectsService) { }
+  constructor(private projectsService: ProjectsService, private modalService: ModalService) { }
 
   ngOnInit() {
 
@@ -33,13 +33,17 @@ export class UsersPanelComponent implements OnInit {
   }
 
   deleteProject() {
-    this.projectsService.deleteProject(this.activeProject);
+    this.modalService.confirm(
+      'Delete map',
+      'Are you sure you want to delete this map?  All associated features and metadata will be deleted. THIS CANNOT BE UNDONE.',
+      ['Cancel', 'Delete']).subscribe( (answer) => {
+        if(answer === 'Delete') {
+          this.projectsService.deleteProject(this.activeProject);
+        }
+    });
   }
 
   addUser() {
-    console.log(this.addUserForm);
     this.projectsService.addUserToProject(this.activeProject, this.addUserForm.get('username').value);
   }
-
-
 }
